Make project updated

+++ make/compile_all.py
@@ -14,6 +14,7 @@ import compile_lib_core
 import compile_lib_gdi
 import compile_lib_osutil
 import compile_lib_screencapture
+import compile_lib_soundcapture
 import compile_os_win_launcher
 import compile_os_win_service
 import compile_os_win_updater
@@ -22,23 +23,36 @@ import compile_os_win_updater
 class CompileAll():
     
     def __init__(self):
-        self._cpp_compiler_flags=None
-        self._linker_flags=None
         self._b32bit=False
     
     def set_32bit(self):
         self._b32bit=True
-        self._cpp_compiler_flags="-m32"
-        self._linker_flags="-m32"
+    
+    def get_path_tmp(self):
+        if self._b32bit:
+            return utils.PATHTMP+"32"
+        else:
+            return utils.PATHTMP
+    
+    def get_path_native(self):
+        if self._b32bit:
+            return utils.PATHNATIVE+"32"
+        else:
+            return utils.PATHNATIVE
     
     def run(self):
         bok=True        
         arstatus=[]
-        utils.init_path(utils.PATHNATIVE)
+        utils.init_path(self.get_path_native())
         utils.info("BEGIN DEPENDENCIES")
         try:
+            if utils.is_windows():
+                self._dependency("lib_gcc", "8.1.0", arstatus)
+                self._dependency("lib_stdcpp", "8.1.0", arstatus)            
             self._dependency("lib_z", "1.2.11", arstatus)
             self._dependency("lib_turbojpeg", "2.0.3", arstatus)
+            self._dependency("lib_opus", "21.3.1", arstatus)
+            self._dependency("lib_rtaudio", "5.1.0", arstatus)
             utils.info("END DEPENDENCIES")
         except:
             bok=False
@@ -55,6 +69,7 @@ class CompileAll():
                     self._compile(compile_os_win_service,arstatus)
                     self._compile(compile_os_win_updater,arstatus)
                 self._compile(compile_lib_screencapture,arstatus)
+                self._compile(compile_lib_soundcapture,arstatus)
                 utils.info("END COMPILE ALL")
             except:
                 bok=False
@@ -77,24 +92,9 @@ class CompileAll():
         mcp = md.Compile()
         smsg=mcp.get_name()
         try:
-            cflgs=self._cpp_compiler_flags
-            if cflgs is not None:
-                if utils.is_windows():        
-                    mcp.set_cpp_compiler_flags("windows",cflgs)                
-                elif utils.is_linux():
-                    mcp.set_cpp_compiler_flags("linux",cflgs)                
-                elif utils.is_mac():
-                    mcp.set_cpp_compiler_flags("mac",cflgs)
-                    
-            lflgs=self._linker_flags
-            if lflgs is not None:
-                if utils.is_windows():        
-                    mcp.set_linker_flags("windows",lflgs)                
-                elif utils.is_linux():
-                    mcp.set_linker_flags("linux",lflgs)                
-                elif utils.is_mac():
-                    mcp.set_linker_flags("mac",lflgs)
-                    
+            if self._b32bit:
+                mcp.set_32bit()
+                  
             mcp.run()
             smsg+=" - OK!"
             ars.append(smsg)
@@ -104,7 +104,7 @@ class CompileAll():
             raise e
     
     def _dependency_post_fix(self,snm,sver):
-        spth=utils.PATHTMP + os.sep + snm;
+        spth=self.get_path_tmp() + os.sep + snm;
         if snm=="lib_z":
             if utils.is_mac():   
                 #CORREGGE zutil.h
@@ -119,7 +119,7 @@ class CompileAll():
                 f.close()
     
     def _dependency(self,snm,sver,ars):        
-        spth=utils.PATHTMP + os.sep + snm;
+        spth=self.get_path_tmp() + os.sep + snm;
         smsg = snm + " " + sver
         utils.info("BEGIN " + snm)        
         try:
@@ -145,10 +145,13 @@ class CompileAll():
                 utils.init_path(spth)
                 utils.info("download headers and library ...")
                 nurl = utils.get_node_url()
-                appnm="headers_" + snm + ".zip"
-                utils.download_file(nurl + "getAgentFile.dw?name=" + appnm , spth + os.sep + appnm)
-                utils.unzip_file(spth + os.sep + appnm, spth + os.sep)
-                utils.remove_file(spth + os.sep + appnm)
+                
+                if snm is not "lib_gcc" and snm is not "lib_stdcpp":
+                    appnm="headers_" + snm + ".zip"
+                    utils.download_file(nurl + "getAgentFile.dw?name=" + appnm , spth + os.sep + appnm)
+                    utils.unzip_file(spth + os.sep + appnm, spth + os.sep)
+                    utils.remove_file(spth + os.sep + appnm)
+                
                 appnm=snm + "_" + sfx + ".zip"
                 utils.download_file(nurl + "getAgentFile.dw?name=" + appnm , spth + os.sep + appnm)
                 utils.unzip_file(spth + os.sep + appnm, spth + os.sep, "native/")
@@ -163,7 +166,7 @@ class CompileAll():
             #COPY LIB TO NATIVE
             for f in os.listdir(spth):
                 if f.endswith('.dll') or f.endswith('.so') or f.endswith('.dylib'): 
-                    shutil.copy2(spth + os.sep + f, utils.PATHNATIVE + os.sep + f)
+                    shutil.copy2(spth + os.sep + f, self.get_path_native() + os.sep + f)
             
             #POST FIX
             self._dependency_post_fix(snm,sver)
